still having problems
