+++ src/index.html
@@ -6,7 +6,7 @@
     <title>Sleep Outside | Home</title>
     <link rel="stylesheet" href="/css/style.css" />
     <script src="js/main.js" type="module"></script>
-    <script src="../js/cart.js" type="module"></script>
+
 
   </head>
   <body>
@@ -33,7 +33,7 @@
           mission is to provide you with everything you need to do this
           comfortably at an affordable price.
         </p>
-        <a href="checkout/index.html">hello</a>
+
       </section>
       <section class="products">
         <h2>Top Products</h2>
