+++ src/js/main.js
@@ -1,4 +1,7 @@
 import { productList } from "./productList.mjs";
+import { loadHeaderFooter, updateCartIconCount } from "./utils.mjs";
 productList(".product-list", "tents");
 
-
+loadHeaderFooter().then(() => {
+  updateCartIconCount(); 
+});
