+++ src/js/product.js
@@ -6,11 +6,13 @@ finish setting everything up. */
 import { getParam } from "./utils.mjs";
 // import { findProductById } from "./productData.mjs";
 import productDetails from "./productDetails.mjs";
-import { loadHeaderFooter}  from "./utils.mjs";
+import {updateCartIconCount } from "./utils.mjs";
 
 const productId = getParam("product");
 productDetails(productId);
-loadHeaderFooter();
+
+updateCartIconCount(); 
+
 
 // function addProductToCart(product) {
 //   let currentCart = getLocalStorage("so-cart") || [];
