+++ src/js/shoppingCart.mjs
@@ -1,0 +1,100 @@
+import { getLocalStorage, setLocalStorage, getCartItemCount } from "./utils.mjs";
+
+export default function shoppingCart() {
+  renderCartContents();
+  updateCartIconCount();
+}
+
+function renderCartContents() {
+  let cartItems = getLocalStorage("so-cart") || [];
+
+  // Safety: ensure array format
+  if (!Array.isArray(cartItems)) {
+    cartItems = cartItems ? [cartItems] : [];
+  }
+
+  // Ensure each item has a quantity (default = 1)
+  cartItems = cartItems.map(item => {
+    if (!item.quantity) item.quantity = 1;
+    return item;
+  });
+
+  // Render items
+  const htmlItems = cartItems.map(cartItemTemplate);
+    const output = document.querySelector(".product-list");
+    output.innerHTML = htmlItems.join("");
+
+  // Set up event listeners
+  document.querySelectorAll(".remove-item").forEach(button => {
+    button.addEventListener("click", (event) => {
+      const productId = event.target.dataset.id;
+      removeFromCart(productId);
+    });
+  });
+
+  document.querySelectorAll(".cart-quantity-input").forEach(input => {
+    input.addEventListener("change", (event) => {
+      const newQty = parseInt(event.target.value);
+      const productId = event.target.dataset.id;
+      if (newQty > 0) {
+        updateCartQuantity(productId, newQty);
+      }
+    });
+  });
+}
+
+function cartItemTemplate(item) {
+  return `<li class="cart-card divider">
+    <a href="#" class="cart-card__image">
+      <img src="${item.Image}" alt="${item.Name}" />
+    </a>
+    <a href="#">
+      <h2 class="card__name">${item.Name}</h2>
+    </a>
+    <p class="cart-card__color">${item.Colors[0].ColorName}</p>
+
+    <p class="cart-quantity">
+      Qty: <input type="number" class="cart-quantity-input" data-id="${item.Id}" 
+      id="quantity-${item.Id}" min="1" value="${item.quantity}" />
+    </p>
+
+    <p class="cart-card__price">$${item.FinalPrice}</p>
+    <button class="remove-item" data-id="${item.Id}">X</button>
+  </li>`;
+}
+
+function updateCartQuantity(productId, newQuantity) {
+  let cart = getLocalStorage("so-cart") || [];
+  const itemIndex = cart.findIndex(item => item.Id === productId);
+  if (itemIndex !== -1) {
+    cart[itemIndex].quantity = newQuantity;
+  }
+  setLocalStorage("so-cart", cart);
+  updateCartIconCount();
+  renderCartContents(); // re-render
+}
+
+function removeFromCart(productId) {
+  let cart = getLocalStorage("so-cart") || [];
+  const itemIndex = cart.findIndex(item => item.Id === productId);
+
+  if (itemIndex !== -1) {
+    if (cart[itemIndex].quantity > 1) {
+      cart[itemIndex].quantity -= 1;
+    } else {
+      cart.splice(itemIndex, 1);
+    }
+  }
+
+  setLocalStorage("so-cart", cart);
+  renderCartContents();
+}
+
+function updateCartIconCount() {
+  const count = getCartItemCount();
+  const cartCountEl = document.getElementById("cart-count");
+  const cartCountPlainEl = document.getElementById("cart-count-plain");
+
+  if (cartCountEl) cartCountEl.textContent = count;
+  if (cartCountPlainEl) cartCountPlainEl.textContent = count;
+}
